pyproject.toml
[project]
name = "pipe-ftl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "click>=8.3.0",
    "playwright>=1.55.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "selenium>=4.36.0",
]
